Correctly switch resource mode on models
