+++ panel/models/ace.py
@@ -4,7 +4,8 @@ Defines custom AcePlot bokeh model to render Ace editor.
 from bokeh.core.properties import String, Override, Dict, Any, List, Bool, Enum
 from bokeh.models import HTMLBox
 
-from ..util import classproperty, bundled_files
+from ..io.resources import bundled_files
+from ..util import classproperty
 from .enums import ace_themes
