+++ panel/models/deckgl.py
@@ -16,7 +16,8 @@ from bokeh.core.properties import (
 )
 from bokeh.models import HTMLBox, ColumnDataSource
 
-from ..util import classproperty, bundled_files
+from ..io.resources import bundled_files
+from ..util import classproperty
 
 
 class DeckGLPlot(HTMLBox):
